CP-54351 move sysdevice from /opt
